refactor: create query language package
